# Implement much faster dot product algorithm for tensors